refactor: 파일 및 함수 이름 변경
